feat: framwork and framework controls
This commit creates the Framework and FrameworkControl models, along with the routes and CRUD templates.

It also includes tests (which will fail for now) and seed data, including the ITILv4 and IO27001:2022 frameworks along with their controls / practices.

In future commits, I'll add a ComplianceLink to all the objects supported in OpsDeck. This will allow associating any policy, asset, document, etc. to any control of any framework.

Next task will be adding a compliance  center where you can check the realtime compliance of the framework controls and export it as a list.

After that will come the concept of audits and remediation plans, which will use a snapshot of the compliance status instead of using the live compliance stats.